% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rucc2003}
\alias{rucc2003}
\title{2003 Rural-Urban Continuum Codes}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3142 rows and 8 columns.}
\source{
https://www.ers.usda.gov/data-products/rural-urban-continuum-codes/
}
\usage{
rucc2003
}
\description{
2003 Rural-Urban Continuum Codes
}
\keyword{datasets}

